+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Add `prop` function with `KProperty1` argument.
+
+### Deprecated
+- Deprecated `prop` function with `KCallable` argument. Use the new overload
+with type-safe `KProperty1` argument or another overload with explicit name and lambda.
+
 ## [0.22] 2020-03-11
 
 ### Added